Separate pinchDragTranslate and pinchTranslate
